Creating an empty bundle if doesn't exist

+++ docker/create.sh
@@ -5,7 +5,12 @@ set -x
 rm -f requirements-lint.txt* bundle.tgz
 
 wget 'https://raw.githubusercontent.com/ROCm/vllm/refs/heads/main/requirements-lint.txt'
-cp ~/Projects/docker_bundle.tgz bundle.tgz
+
+if [[ -e ~/Projects/docker_bundle.tgz ]] ; then
+    cp ~/Projects/docker_bundle.tgz bundle.tgz
+else
+    tar czvf bundle.tgz --files-from=/dev/null
+fi
 
 docker build -f Dockerfile.vllm \
     --build-arg "UID=$(id -u)" \
